starter/requirements.txt
alembic==1.4.0
azure-common==1.1.24
azure-nspkg==3.0.2
azure-storage==0.36.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.8
Flask==0.12.5
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.1
MarkupSafe==1.1.1
pip==10.0.1
pycparser==2.19
pyodbc==4.0.28
python-dateutil==2.8.1
python-editor==1.0.4
requests==2.22.0
setuptools==39.0.1
six==1.14.0
SQLAlchemy==1.3.13
urllib3==1.25.8
Werkzeug==0.16.1
WTForms==2.2.1
